requirements.txt
future==0.18.2
Metaphone==0.6
nltk==3.4.5
pronounceable==0.1.3
python-whois==0.7.2
PyYAML==5.3
six==1.14.0
